ulogd-NFCT-make-cache-handling-smarter.diff
The previous cache handling approach was independant on the number of
hash buckets and always ran over 1/16 of the hash buckets, which lead
to SNDBUF overrun on some systems with many longstanding connections.

The new approach is to go over 16 buckets at once, independant on the
number of hash buckets.  This also leads to the fact that increasing
the number of hash buckets actually may solve future problems.

Signed-off-by: Holger Eitzenberger <heitzenberger@astaro.com>