ulogd-FIX-Do-ROLLBACK-in-case-of-SQLITE_BUSY.diff
[FIX] Do ROLLBACK in case of SQLITE_BUSY

As described in

 http://www.sqlite.org/cvstrac/wiki?p=CorruptionFollowingBusyError

SQLITE 3.3.8 contains an error when transactions other than EXCLUSIVE
are used.  If a SQL commands to create a transaction or inside
a transaction ever gets SQLITE_BUSY a ROLLBACK has to be done, as
otherwise a corrupt database might be the result.

This fix does a ROLLBACK in all cases.

Signed-off-by: Holger Eitzenberger <heitzenberger@astaro.com>