DB: conditionally remove commit timer on stop()
If a DB plugin is stopped if not started correctly the DB commit timer
is removed unconditionally, leading to a crash or to a failed assertion
(if assertions enabled).

The fix is to remove the assertion, as ulogd_upi_stop() happens in
many different places at various times and also to remove the timer
only conditionally.

Signed-off-by: Holger Eitzenberger <holger@eitzenberger.org>