NFCT: fix logic on _NEW event
It is quite possible for a new event to arrive when there is still a
old conntrack in one of the caches, e. g. an ICMP echo request if the
previous echo reply was missed.

Therefore remove the assertion and add the logic to handle this case
correctly by using the newly created conntrack object as well as by
updating the time on it.

A future improvement might be to not to replace the conntrack but to
selectively update some fields on the old one, e. g. the counters to
reflect the old traffic plus the new one.

Signed-off-by: Holger Eitzenberger <holger@eitzenberger.org>