Fix crash on reconfigure in plugins which were not started
This was due to ulogd_unregister_fd() happening in instance->stop()
unconditionally despite the fact that the ulogd_fd wasn't properly
registered, resulting in a segmentation fault when removing from the
ulogd_fd list.

Solve this issue by adding the ulogd_init_fd() function to the core
and using it in the instance->configure() handlers, among others to
properly initialize the llist_head.  Note that it's specifically not
sufficient to add it to instance->start(), because a start might even
not happened before.

Signed-off-by: Holger Eitzenberger <holger@eitzenberger.org>