Always clear source links
There was a subtle bug in the state handling logic which resulted in
the fact that on a reconfigure of a stack only some of the input keys
of a plugin instance were reset, namely those which

 * which had PF_RECONF set, see stack_reconfigure()
 * had a stop() routine

If there was a plugin in a stack which lacked one of those ulogd failed
at reconfigure (e. g. SIGHUP) with a message like

 ulogd[19133]: input key `oob.prefix' already has source

This patch fixes that by doing the reset of the source links always.
I therefore move the PF_RECONF check from stack_reconfigure() to
ulogd_upi_stop() and rearrange ulogd_upi_stop() to do the reset always.

Signed-off-by: Holger Eitzenberger <holger@eitzenberger.org>