Turn config_key.u.str into a pointer
config_key.u.str is a string array with size 255, which is
rarely used completely and therefore mostly unused.

Therefore turn it into a character pointer and allocate a
string just long enough during initialization.

Note though that there is currently no code to free that memory,
but this is ok as long as the config file is only parsed at
start of ulogd.

Signed-off-by: Holger Eitzenberger <holger@eitzenberger.org>