Fix key reset
In the commit which introduced ulogd_value in order to be used
by the database layer, the key.type field moved into the newly
created ulogd_value.  However, this broke the key reset in
stack_clean_results(), which just overwrites everything in the
key value, overwriting the type info.  Subsequent key accesses
(get or set) then go BUG() like this:

 Oct  7 17:33:10 kruemel ulogd: ip.protocol: invalid key set (0 <-> 5)
 Oct  7 17:33:10 kruemel ulogd: BUG at plugin.c:886

The fix is to reset just the key flag, which should be
sufficient.  Also gives a small performance improvement.

Signed-off-by: Holger Eitzenberger <holger@eitzenberger.org>