Fix key check by removing key.source overlay
The key check failed after introducing ulogd_key and more
importantly by moving key.type to ulogd_value.  Because of
the overlay of key.u.source and key.u.val this broke the
key check in stack_resolve_keys(), because the value.type
was already set then.

Fix it by removing this overlay (remove that union).  This
increases the ulogd_key size by 4 bytes, which is not a
problem at all.

Signed-off-by: Holger Eitzenberger <holger@eitzenberger.org>