Reset key source pointer on error
Also it is specifically not an error if a key pointer already has
a source link.  This may happen e. g. if a database plugin
errors out in it's ->start() routine, having already the source
links set.  This is what happens:

  PGSQL: configure
  NFLOG: configure
  stack_resolve_keys(stack);
  PGSQL: start()             	/* error, back to PsInit */
  PGSQL: configure()
  stack_resolve_keys(stack);	/* key->source already set */

Fix it by removing the check in stack_resolve_keys(), also
reset key->source correctly on other errors.

Signed-off-by: Holger Eitzenberger <holger@eitzenberger.org>